# swagger.yaml
openapi: 3.0.3
info:
  title: API Biblioteca Digital
  version: 1.0.0
  description: API para consulta e gerenciamento de livros
servers:
  - url: http://localhost:3000
paths:
...
  /emprestimos:
    get:
      tags:
        - Empréstimos
      summary: Lista todos os empréstimos
      responses:
        '200':
          description: Lista de empréstimos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Emprestimo'
    post:
      tags:
        - Empréstimos
      summary: Cria um novo empréstimo
      description: |
        Cria um novo empréstimo verificando as seguintes regras:
        - Usuário não pode ter mais de 3 empréstimos simultâneos
        - Usuário não pode ter empréstimos em atraso (mais de 15 dias)
        - Livro deve estar disponível
        - Prazo de devolução é de 15 dias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmprestimoInput'
      responses:
        '201':
          description: Empréstimo criado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Emprestimo'
                  - type: object
                    properties:
                      mensagem:
                        type: string
                        example: 'Empréstimo realizado com sucesso'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    enum:
                      - 'Usuário e livro são obrigatórios'
                      - 'Usuário possui empréstimos em atraso'
                      - 'Usuário já atingiu o limite de 3 empréstimos simultâneos'
                      - 'Livro não disponível para empréstimo'
                  codigo:
                    type: integer
                    example: 400
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Recurso não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                    enum:
                      - 'Usuário não encontrado'
                      - 'Livro não encontrado'
                  codigo:
                    type: integer
                    example: 404
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro:
                    type: string
                  codigo:
                    type: integer
                    example: 500
                  timestamp:
                    type: string
                    format: date-time
...
  /emprestimos/{id}/devolucao:
    put:
      tags:
        - Empréstimos
      summary: Devolve um livro emprestado
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Livro devolvido com sucesso
        '404':
          description: Empréstimo não encontrado
        '500':
          description: Erro interno ao devolver empréstimo
...
  /livros/{id}:
    get:
      tags:
        - Livros
      summary: Obtém um livro pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Livro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Livro'
        '404':
          description: Livro não encontrado
    put:
      tags:
        - Livros
      summary: Atualiza um livro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LivroInput'
      responses:
        '200':
          description: Livro atualizado
    delete:
      tags:
        - Livros
      summary: Remove um livro
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Livro removido
...
components:
  schemas:
...
    Emprestimo:
      type: object
      properties:
        id:
          type: integer
        usuario_id:
          type: integer
        livro_id:
          type: integer
        data_emprestimo:
          type: string
          format: date
        data_devolucao:
          type: string
          format: date
        devolvido:
          type: boolean
    EmprestimoInput:
      type: object
      required:
        - usuario_id
        - livro_id
      properties:
        usuario_id:
          type: integer
        livro_id:
          type: integer
    Autor:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
...
    Livro:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        isbn:
          type: string
        ano_publicacao:
          type: integer
        disponivel:
          type: boolean
        autor:
          $ref: '#/components/schemas/Autor'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    LivroInput:
      type: object
      required:
        - titulo
        - isbn
        - ano_publicacao
        - autor_id
      properties:
        titulo:
          type: string
        isbn:
          type: string
        ano_publicacao:
          type: integer
        disponivel:
          type: boolean
        autor_id:
          type: integer
    Link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
        method:
          type: string